+++ app/Http/Controllers/TagsController.php
@@ -88,6 +88,11 @@ class TagsController extends Controller
      */
     public function destroy(Tag $tag)
     {
+        if ($tag->posts->count()>0) {
+            session()->flash('error','Cannot delete this tag because it has an active article.');
+            return redirect()->back();
+        }
+
         $tag->delete();
         session()->flash('success','Deleted Tag successfully');
         return redirect(route('tags.index'));
